% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_name_or_code.R
\name{country_name_or_code}
\alias{country_name_or_code}
\title{Convert between country codes and country names}
\usage{
country_name_or_code(x)
}
\arguments{
\item{x}{A character vector containing either 2-letter country codes or country
names.}
}
\value{
A character vector of the same length as \code{x}, containing either
country names or 2-letter country codes, depending on the input.
}
\description{
This function converts between 2-letter country codes and country names. If the
input contains all 2-letter codes, the function returns the corresponding country
names. If the input contains all country names, the function returns the
corresponding 2-letter codes.
}
\examples{
# Convert country codes to country names
country_name_or_code(c("US", "GB", "CA"))

# Convert country names to country codes
country_name_or_code(c("United States", "United Kingdom", "Canada"))

# Error: input contains both country codes and country names
country_name_or_code(c("US", "GB", "CA"))

}
