% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_clicks.R
\name{highlight_clicks}
\alias{highlight_clicks}
\title{Highlight Plotly Points}
\usage{
highlight_clicks(plotly_obj, layer_brush = 0)
}
\arguments{
\item{plotly_obj}{plotly object with points.}

\item{layer_brush}{Provide a numeric value. If none supplied, layer_brush = 0 will be used.
In shiny, if this value is larger than 3, consider using modules.}
}
\value{
plotly_object
}
\description{
This JavaScript code defines a function that is used in a Plotly chart to capture changes in
the visibility of the traces on the chart.
The layer_brush is used to extract a DOM element from a D3.js selection to access
its properties and modify its attributes.
}
\examples{
## Not run:
library(dplyr)
library(ggplot2)

p_1 <- mtcars \%>\%
  ggplot(aes(x = mpg, y = disp)) +
  geom_point()

library(plotly)
library(htmlwidgets)

p_1 \%>\%
 ggplotly() \%>\%
 onRender(jsCode = highlight_clicks(.))

## End(Not run)
}
