% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marchenko_pastur_pdf.R
\name{marchenko_pastur_pdf}
\alias{marchenko_pastur_pdf}
\title{Calculate the Marcenko-Pastur Probability Density Function (PDF)}
\usage{
marchenko_pastur_pdf(variance, q, pts)
}
\arguments{
\item{variance}{The variance value for the Marcenko-Pastur distribution.}

\item{q}{The ratio of variables to observations.}

\item{pts}{The number of points to use in the PDF.}
}
\value{
A data.frame containing the Marcenko-Pastur PDF values and their
corresponding index values.
}
\description{
Given a variance value, q, and number of points, calculates the PDF for the
Marcenko-Pastur distribution. The q ratio is explicit for clarity (for me).
}
\examples{
pdf <- marchenko_pastur_pdf(variance = 1, q = 0.5, pts = 100)
plot(pdf$index, pdf$pdf, type = "o")
}
\seealso{
\url{https://en.wikipedia.org/wiki/Marchenko\%E2\%80\%93Pastur_distribution#Application_to_correlation_matrices}
}
