% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_manipulations.R
\name{plot_fct_lump}
\alias{plot_fct_lump}
\title{Plot a categorical variable by lumping infrequent levels}
\usage{
plot_fct_lump(data, col_name, min_val)
}
\arguments{
\item{data}{A data frame containing the column to be plotted.}

\item{col_name}{A character string specifying the column name to be plotted.}

\item{min_val}{An integer specifying the minimum frequency count for a level to be kept in the plot.}
}
\value{
A bar plot of the specified column, with relabeled and lumped infrequent levels.
}
\description{
Returns a bar plot of the frequency distribution of the specified column.
}
\examples{
library(ggplot2)
library(dplyr)
library(forcats)
starwars |> plot_fct_lump("skin_color", 3)
}
