% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_manipulations.R
\name{summarize_diss}
\alias{summarize_diss}
\title{Summarize data frame with target summary statistics}
\usage{
summarize_diss(data, summary_vars, summary_stats)
}
\arguments{
\item{data}{A data frame to summarize.}

\item{summary_vars}{A set of columns of interest.}

\item{summary_stats}{A set of summary statistics of interest.}
}
\value{
Summarized object with groups intact.
}
\description{
summarize_diss takes a data frame to summarize data in a concise format.
The output closely resembles skimr::skim(), but with all the unnecessary information removed.
}
\examples{
library(dplyr)
library(tidyr)
library(purrr)
mtcars \%>\%
 group_by(cyl, gear) \%>\%
 summarize_diss(c("mpg", "disp"), c("mean", "sd", "median", "min", "max"))

}
