% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_manipulations.R
\name{summarize_enkeyed}
\alias{summarize_enkeyed}
\title{Summarize a numeric vector by quantiles within groups}
\usage{
summarize_enkeyed(x, probs = c(0.25, 0.5, 0.75), ...)
}
\arguments{
\item{x}{A numeric vector to be summarized by quantiles}

\item{probs}{A numeric vector of quantile probabilities to compute.
Defaults to c(0.25, 0.5, 0.75).}

\item{...}{Additional arguments to be passed to the quantile() function
(e.g., na.rm = TRUE).}
}
\value{
A list containing a data frame with the quantile values and their
corresponding probabilities.
}
\description{
This function summarizes a numeric vector by quantiles within groups,
returning a data frame with the quantile values and their corresponding
probabilities.
}
\examples{
library(dplyr)
library(tidyr)
iris \%>\%
  group_by(Species)  |>
  summarise(pl_qtile = summarize_enkeyed(Petal.Length)) |>
  unnest(pl_qtile)
}
\seealso{
\href{https://speakerdeck.com/jennybc/row-oriented-workflows-in-r-with-the-tidyverse}{Jenny Bryan's talk}.
}
