% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisean_routines.R
\name{tisean_diss}
\alias{tisean_diss}
\title{Call TISEAN routines.}
\usage{
tisean_diss(
  tisean_path = "/mnt/working_volume/tisean/TISEAN_3.0.1_unix/Tisean_3.0.1",
  time_series = NULL,
  routine = NULL,
  remove_header = FALSE,
  show_output_on_console = FALSE,
  ...
)
}
\arguments{
\item{tisean_path}{A string indicating the path to the Tisean directory}

\item{time_series}{A matrix or data frame containing the time series data.}

\item{routine}{A string indicating the Tisean routine to use.}

\item{remove_header}{A logical value indicating whether to remove the header from the output. Default is FALSE.}

\item{show_output_on_console}{A logical value indicating whether to show the output on console. Default is FALSE.}

\item{...}{Additional parameters to be passed to the Tisean routine.}
}
\value{
A numeric vector or matrix.
}
\description{
A verbose method before I figure out how to do this in rcpp!
}
\examples{
\dontrun{
library(dissthispkgr)
library(dplyr)
henon_m <- tisean_diss(
  routine = "henon",
  remove_header = FALSE,
  show.output.on.console = FALSE, l = 10000)
plot(henon_m)
}
}
