% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zernike_expansion}
\alias{zernike_expansion}
\title{Zernike expansion in Rcpp}
\usage{
zernike_expansion(x, y, noll, normalized = TRUE)
}
\arguments{
\item{x}{NumericVector of x-coordinates.}

\item{y}{NumericVector of y-coordinates.}

\item{noll}{Integer specifying the Noll index of the Zernike polynomial.}

\item{normalized}{Boolean indicating whether the expansion should be normalized.}
}
\value{
NumericVector of the Zernike expansion coefficients.
}
\description{
Calculates the Zernike expansion of a set of points in Cartesian coordinates.
This is a slightly faster version of the R code I wrote a while ago. The esoteric types like
Bhatia-Wolf will be added later.
}
\examples{
x <- c(1, 2, 3)
y <- c(4, 5, 6)
zernike_expansion(x, y, 3, TRUE)

}
