% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_spec_diss.R
\name{fourier_spec_diss}
\alias{fourier_spec_diss}
\title{Plot Fourier spectrogram of a signal}
\usage{
fourier_spec_diss(tbl, title = "signal")
}
\arguments{
\item{tbl}{A data frame containing the time and signal columns}

\item{title}{A character string specifying the title of the spectrogram (default is "signal")}
}
\value{
A ggplot object representing the Fourier spectrogram of the signal
}
\description{
My aim in writing this is to make the TF signal stand out and pleasing.
}
\examples{
library(ggplot2)
library(dplyr)
library(tidyr)

# Create a sine wave with frequency of 100 Hz and amplitude of 1
time <- seq(0, 1000, length.out = 1000)
signal <- sin(2 * pi * 100 * time)

# Add some noise to the signal
set.seed(123)
signal <- signal + rnorm(length(signal), sd = 0.01)

df <- data.frame(time = time, signal = signal)

plot(df, type = "o")
fourier_spec_diss(df)

}
