% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_clicks.R
\name{highlight_clicks}
\alias{highlight_clicks}
\title{Highlight Plotly Points}
\usage{
highlight_clicks(layer_brush = 0)
}
\arguments{
\item{layer_brush}{Provide a numeric value. If none supplied, layer_brush = 0 will be used.}
}
\description{
Highlight Plotly Points
}
